+++ max_div/solver/_strategies/_optimization/_base.py
@@ -3,7 +3,11 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from typing import Any, Self
 
+import numpy as np
+from numpy.typing import NDArray
+
 from max_div.internal.utils import ALMOST_ONE_F32
+from max_div.sampling.poisson import sample_truncated_poisson
 from max_div.solver._scheduling import ParameterSchedule, _evaluate_schedules, _schedules_to_2d_numpy_array
 from max_div.solver._solver_state import SolverState
 from max_div.solver._strategies._base import StrategyBase
@@ -98,7 +102,138 @@ class OptimizationStrategy(StrategyBase, ABC):
     #  Factory Methods
     # -------------------------------------------------------------------------
     @classmethod
-    def dummy(cls) -> Self:
-        from ._optim_dummy import OptimDummy
+    def random_swaps(cls) -> Self:
+        from ._optim_random_swaps import OptimRandomSwaps
 
-        return OptimDummy()
+        return OptimRandomSwaps()
+
+
+# =================================================================================================
+#  Swap-Based Optimization Strategy base class
+# =================================================================================================
+class SwapBasedOptimizationStrategy(OptimizationStrategy, ABC):
+    """
+    Base class for swap-based optimization strategies, where in each iteration 'n' items are removed from the current
+    selection and replaced by 'n' new items, but only if the swap improves the overall score.
+
+    n is sampled from a truncated Poisson distribution with range [min_swap_size, max_swap_size] and lambda
+    parameter swap_size_lambda, the latter of which can be set to a ParameterSchedule or a fixed value.
+
+    Optionally constraint scores can be treated as soft constraints, in which case diversity score is mixed in with
+    the constraint score to a certain degree.  This parameter can also be scheduled.
+
+    Child classes need to implement the way in which...
+      - n items are selected for removal from the current selection
+      - n new items are selected for addition to the current selection
+    """
+
+    # -------------------------------------------------------------------------
+    #  Constructor
+    # -------------------------------------------------------------------------
+    def __init__(
+        self,
+        name: str | None = None,
+        min_swap_size: int = 1,
+        max_swap_size: int = 1,
+        swap_size_lambda: float | ParameterSchedule = 1.0,
+        constraint_softness: float | ParameterSchedule = 0.0,
+        scheduled_params: dict[str, Any] | None = None,
+    ):
+        super().__init__(
+            name=name,
+            scheduled_params=(scheduled_params or dict())
+            | dict(
+                swap_size_lambda=swap_size_lambda,
+                constraint_softness=constraint_softness,
+            ),
+        )
+        self.min_swap_size: np.int32 = np.int32(min_swap_size)
+        self.max_swap_size: np.int32 = np.int32(max_swap_size)
+        self.swap_size_lambda = swap_size_lambda
+        self.constraint_softness = constraint_softness
+
+    # -------------------------------------------------------------------------
+    #  Single Iteration
+    # -------------------------------------------------------------------------
+    def _perform_single_iteration(self, state: SolverState, progress_frac: float):
+        """
+        Perform one iteration of the swap-based optimization strategy:
+          - determine swap size n
+          - remove n samples from current selection
+          - add n new samples to current selection
+          - if this swap did not improve the score -> revert to previous selection
+        """
+
+        # --- prep ---
+        seed = self.next_seed()  # we can use the same seed for all sampling in this iteration
+
+        # --- determine swap size ---
+        n = sample_truncated_poisson(
+            self.min_swap_size,
+            self.max_swap_size,
+            np.float32(self.swap_size_lambda),
+            seed=seed,
+        )
+
+        # --- take snapshot & remove n samples ---
+
+        # snapshot
+        state.set_snapshot()
+        score_before = state.score
+
+        # add
+        samples_to_remove = self._samples_to_be_removed(state, n, seed)
+        for s in samples_to_remove:
+            state.remove(s)
+
+        # --- add n samples and evaluate ---
+
+        # add
+        samples_to_add = self._samples_to_be_added(state, n, samples_just_removed=samples_to_remove, seed=seed)
+        for s in samples_to_add:
+            state.add(s)
+
+        # evaluate
+        score_after = state.score
+        if score_after.as_tuple(self.constraint_softness) <= score_before.as_tuple(self.constraint_softness):
+            state.restore_snapshot()  # score didn't improve -> restore snapshot
+
+    # -------------------------------------------------------------------------
+    #  Abstract methods
+    # -------------------------------------------------------------------------
+    @abstractmethod
+    def _samples_to_be_removed(
+        self,
+        state: SolverState,
+        n: np.int32,
+        seed: np.int64,
+    ) -> NDArray[np.int32]:
+        """
+        Determine which n samples to remove from the current selection.  The values returned should be indices present
+        in state.selected_index_array.
+        :param state: (SolverState) current solver state, with # selected samples = k
+        :param n: (np.int32) number of samples to be removed  (swap size)
+        :param seed: (np.int64) random seed to be used for any sampling
+        :return: (int32 ndarray) of shape (n,) with indices of samples to be REMOVED
+        """
+        raise NotImplementedError()
+
+    @abstractmethod
+    def _samples_to_be_added(
+        self,
+        state: SolverState,
+        n: np.int32,
+        samples_just_removed: NDArray[np.int32],
+        seed: np.int64,
+    ) -> NDArray[np.int32]:
+        """
+        Determine which n samples to add to the current selection, right after having removed n samples.
+        The values returned should be indices present in state.not_selected_index_array.
+        :param state: (SolverState) current solver state, with # selected samples = k-n
+        :param n: (np.int32) number of samples to be added  (swap size)
+        :param samples_just_removed: (int32 ndarray) of shape (n,) with indices of samples that were just removed
+                                          (potentially to be taken into account to avoid resampling them)
+        :param seed: (np.int64) random seed to be used for any sampling
+        :return: (int32 ndarray) of shape (n,) with indices of samples to be ADDED
+        """
+        raise NotImplementedError()