+++ tests/solver/_strategies/_optimization/test_optim_random_swaps.py
@@ -1,0 +1,54 @@
+import pytest
+
+from max_div.benchmarks import BenchmarkProblemFactory
+from max_div.solver import DiversityMetric, MaxDivProblem
+from max_div.solver._solver_state import SolverState
+from max_div.solver._strategies._optimization._optim_random_swaps import OptimRandomSwaps
+
+
+@pytest.mark.parametrize("size", [1, 2, 10])
+@pytest.mark.parametrize("problem_name", ["A1", "A2", "A3", "A4", "A5"])
+def test_optim_random_swaps(problem_name: str, size: int):
+    """
+    Test OptimRandomSwaps strategy on reference problems, with very rudimentary initialization,
+    to see if we're optimizing.
+    """
+
+    # --- arrange -----------------------------------------
+
+    # prepare problem & solver state
+    problem: MaxDivProblem = BenchmarkProblemFactory.construct_problem(
+        name=problem_name,
+        size=size,
+        diversity_metric=DiversityMetric.approx_geomean_separation(),
+    )
+    solver_state = SolverState.new(
+        vectors=problem.vectors,
+        k=problem.k,
+        distance_metric=problem.distance_metric,
+        diversity_metric=problem.diversity_metric,
+        diversity_tie_breakers=[],
+        constraints=problem.constraints,
+    )
+
+    # initialize solver state
+    for idx in range(problem.k):
+        solver_state.add(idx)
+
+    # prepare strategy
+    strategy: OptimRandomSwaps = OptimRandomSwaps()
+    strategy.seed = 42
+    initial_score = solver_state.score
+    n_iterations = 100
+
+    # --- act ---------------------------------------------
+    strategy.perform_n_iterations(
+        state=solver_state,
+        n_iters=n_iterations,
+        current_progress_frac=0.0,
+        progress_frac_per_iter=1e-6,  # dummy value
+    )
+
+    # --- assert ------------------------------------------
+    assert len(solver_state.selected_index_array) == problem.k, "Number of selected items should remain k."
+    assert solver_state.score > initial_score, "We should be optimizing"